# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blood Strike Scrim Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scrims/{id}/matches:
    get:
      operationId: getScrimMatches
      tags: [scrims]
      summary: Get matches for a scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a match in a scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchBody"
      responses:
        "201":
          description: Created match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchBody"
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}/scores:
    get:
      operationId: getMatchScores
      tags: [scoreboard]
      summary: Get scores for a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
    post:
      operationId: submitScore
      tags: [scoreboard]
      summary: Submit score for a team in a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoreBody"
      responses:
        "201":
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
  /scrims/{id}/scoreboard:
    get:
      operationId: getScrimScoreboard
      tags: [scoreboard]
      summary: Get full scoreboard for a scrim (ranked)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranked scoreboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreboardEntry"
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnnouncementBody"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Match:
      type: object
      properties:
        id:
          type: integer
        scrimId:
          type: integer
        matchNumber:
          type: integer
        mapName:
          type: string
        roundNumber:
          type: integer
        status:
          type: string
          enum: [pending, active, finished]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - scrimId
        - matchNumber
        - mapName
        - roundNumber
        - status
        - startTime
        - endTime
        - createdAt
    CreateMatchBody:
      type: object
      properties:
        matchNumber:
          type: integer
        mapName:
          type: string
        roundNumber:
          type: integer
      required:
        - matchNumber
        - mapName
        - roundNumber
    UpdateMatchBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, finished]
        startTime:
          type: string
        endTime:
          type: string
    Score:
      type: object
      properties:
        id:
          type: integer
        matchId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        kills:
          type: integer
        placement:
          type: integer
        placementPoints:
          type: integer
        killPoints:
          type: integer
        violationDeductions:
          type: integer
        totalPoints:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - matchId
        - teamId
        - teamName
        - kills
        - placement
        - placementPoints
        - killPoints
        - violationDeductions
        - totalPoints
        - createdAt
    SubmitScoreBody:
      type: object
      properties:
        teamId:
          type: integer
        kills:
          type: integer
        placement:
          type: integer
      required:
        - teamId
        - kills
        - placement
    ScoreboardEntry:
      type: object
      properties:
        rank:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        totalKills:
          type: integer
        totalPoints:
          type: integer
        matchesPlayed:
          type: integer
        violations:
          type: integer
      required:
        - rank
        - teamId
        - teamName
        - totalKills
        - totalPoints
        - matchesPlayed
        - violations
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [tournament, event, scrim, update, general]
        isPinned:
          type: boolean
        createdByUsername:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - category
        - isPinned
        - createdByUsername
        - createdAt
        - updatedAt
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [tournament, event, scrim, update, general]
        isPinned:
          type: boolean
        createdBy:
          type: integer
      required:
        - title
        - content
        - category
        - createdBy
    UpdateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [tournament, event, scrim, update, general]
        isPinned:
          type: boolean
...