# lib/api-zod/src/generated/types/scoreboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Strike Scrim Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreboardEntry {
  rank: number;
  teamId: number;
  teamName: string;
  totalKills: number;
  totalPoints: number;
  matchesPlayed: number;
  violations: number;
}