# lib/api-zod/src/generated/types/updateAnnouncementBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Strike Scrim Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAnnouncementBodyCategory } from "./updateAnnouncementBodyCategory";

export interface UpdateAnnouncementBody {
  title?: string;
  content?: string;
  category?: UpdateAnnouncementBodyCategory;
  isPinned?: boolean;
}